feat: create location validator

+++ src/data/services/Project/DeleteProject.service.js
@@ -1,8 +1,12 @@
-import { ProjectRepository } from '../../database/repositories';
+import {
+  ProjectRepository,
+  LocationRepository,
+} from '../../database/repositories';
 
 export class DeleteProjectService {
   async execute({ id_project }) {
     const repository = new ProjectRepository();
+    const locationRepository = new LocationRepository();
 
     const verifyAgencyExists = await repository.findProjectById({
       id_project,
@@ -12,6 +16,17 @@ export class DeleteProjectService {
       return {
         error: `Não há nenhum projeto registrado com este ID -> ${id_project}.`,
       };
+
+    const verifyFk = await locationRepository.verifyProject({
+      id_project,
+    });
+
+    if (verifyFk.length > 0) {
+      return {
+        error:
+          'Não foi possível excluir o Projeto pois existem Localização de Canteiros associadas.',
+      };
+    }
 
     await repository.deleteProject({
       id_project,
